Import shared code
